don't throw in File::~File if CFITSIO returns an error.
Procrastination in handling cfitsio::fits_file_close properly in the
destructor finally bit me when ~File threw inside of a stack unwind
from an exception.  Classic case of why not to throw.

There's a new File::close method which throws.  If it is not called,
the FITS file is closed in File::~File, logging a message to std::cerr if
it fails.  Eventually this should be passed to a logging object.