kludge to work around feature in boost::dynamic_bitset with non-multiple of sizeof(char) number of bits
if bitset.size() is not an integral multiplier of
BitSet::bits_per_block, boost::from_block_range() will corrupt the
bitset and ~dynamic_bitset will abort after sanity checks have failed.

Kludge around this by temporarily inflating the size of the bitset to
be equal to the maximum number of bits that will fit in the minimum
required number of blocks prior to reading, and then resetting it to
the desired number of bits afterwards.