bug fix: copying no columns via Table::copy_columns resulted in extending the table
Table::copy_columns tried to extend the destination table even if no
columns were to be copied.  If the destination table had no columns,
CFITSIO returned an error.  In the case it did have columns, the table
would be extended for no good resaon.

Simply do nothing if there are no columns to copy.