don't use resetHDU to set and then restore CHDU. it can kill performance.
If the CHDU is X, and an operation requires HDU Y, setting CHDU to Y,
performing the operation and resetting to X causes CFITSIO to re-read the
header for both Y and X for each operation, which can destroy performance if
that operation is performed a lot (e.g. row.read()).  One could avoid this
by setting the CHDU to Y prior to the operations being performed, but this
is a subtle point that can be lost on the user; after all, the whole point
of this library is to make so that the user doesn't need to track that stuff.

Instead, just set the HDU to Y (which is a no-op after the first in a
sequence of operations on Y).  This doesn't retain the previous state
of the CHDU, but that should be transparent to the user.