add support for boost::container::vector
Because std::vector<bool> doesn't behave like a real vector, need to
support alternate vector implentations for bool.

To that end:

Column is now just the visible interface layer; all Column classes are formed
via specializations using either ColumnInit or ColumnVector as their
base.


ColumnBase now contains the other metadata which used to be in Column.
ColumnInit is a very thin templated shim which provides an init method
in the constructor which can be overriden by type. This is now how the
specialization of a byte_t array/vector to store bit fields is handled
.

ColumnVector now forms the base class for vector based storage.

Previously everything was built upon Column, which caused cross-talk
when specializations using ColumnVector used it a base. Specializing
Column with a class based on Column caused instantiation of the base
for the type to be handled by the specialization. (Or something like that).

All of the Column* class templates use an init() function in the
constructor which can be overridden by template specialization.

Along the way, it was recognized that the "optimization" of having
CFITSIO write/read directly into/from an array of bools if
sizeof(bool) == sizeof(byte_t) is wrong, as the implementation of a
bool value is undefined.  Each value has to be explicitly converted
(i.e., transferred from byte_t to bool or back).  This resulted in
removing the test in Column<bool> for the "optimization".